+++ OpenccNet/CHANGELOG.md
@@ -9,6 +9,10 @@ the [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
 ## [1.6.1] - Unreleased
 
+### Add
+
+- CLI: convert - added feature `--custom-dict` to enable custom conversion dictionary slot.
+
 ### Changed
 
 - CLI: Optimized `OpenccNet office` subcommand.